This is the Demised Fifth Runner SC Script
This is the Demised Fifth Runner SupperCollider Synth Script.
It plays a little Melody created from the Circles of Fifths. But the Fifths are here a little bit demised. So the Firths here are 20 to 30 Cents flatter than the perfect Fifth.
I hope somebody will like this little Synth and plays a little bit with it.